// src/pipedrive/dtos/deal-dto.ts
import { IOrgDTO } from './org-dto';
import { IPersonDTO } from './person-dto';
import { IUserDTO } from './user-dto';

export interface IDealDTO {
  id: number;
  creator_user_id: IUserDTO;
  user_id: IUserDTO;
  person_id: IPersonDTO;
  org_id: IOrgDTO;
  stage_id: number;
  title: string;
  value: number;
  currency: string;
  add_time: string;
  update_time: string;
  stage_change_time: any;
  active: boolean;
  deleted: boolean;
  status: string;
  probability: any;
  next_activity_date: string;
  next_activity_time: any;
  next_activity_id: number;
  last_activity_id: any;
  last_activity_date: any;
  lost_reason: any;
  visible_to: string;
  close_time: any;
  pipeline_id: number;
  won_time: any;
  first_won_time: any;
  lost_time: any;
  products_count: number;
  files_count: number;
  notes_count: number;
  followers_count: number;
  email_messages_count: number;
  activities_count: number;
  done_activities_count: number;
  undone_activities_count: number;
  participants_count: number;
  expected_close_date: string;
  last_incoming_mail_time: any;
  last_outgoing_mail_time: any;
  label: any;
  renewal_type: string;
  stage_order_nr: number;
  person_name: string;
  org_name: string;
  next_activity_subject: string;
  next_activity_type: string;
  next_activity_duration: any;
  next_activity_note: any;
  group_id: any;
  group_name: any;
  formatted_value: string;
  weighted_value: number;
  formatted_weighted_value: string;
  weighted_value_currency: string;
  rotten_time: any;
  owner_name: string;
  cc_email: string;
  org_hidden: boolean;
  person_hidden: boolean;
}